requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
tzdata>=2024.1
